feat(ext): add lavender module with extended named colours
- fix(ext): correct example assertions in brown and indigo modules
